PuntoSpark-Kafka/requirements.txt
kafka-python==2.0.2
numpy==1.21.4
pandas==1.3.4
py4j==0.10.9.2
pyspark==3.2.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
